+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Upcoming
 
+* Improve boost overview (now shows how much sp need to get invested to make a place boostable and safe)
+
 # 0.6.0
 * Include advanced tactics in calculated attack and defense bonus
 * Show tooltips for resources in plunder screen in local (server) language ([#12](https://github.com/veger/foei/issues/12))