+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Upcoming
 
+# 0.7.0
 * Improve boost overview (now shows how much sp need to get invested to make a place boostable and safe)
 * Renamed 'rewards' to 'incidents'
 * Prevent errors about undefined objects ([#22](https://github.com/veger/foei/issues/22))